refactor: unify epoch_observer impls into a common module
