+++ crypto-ffi/src/generic/mod.rs
@@ -1,4 +1,3 @@
 use crate::{Ciphersuite, Ciphersuites, ClientId, CoreCrypto, CoreCryptoError, CoreCryptoResult, CredentialType};
 
 pub mod context;
-mod epoch_observer;
