+++ crypto-ffi/src/wasm/mod.rs
@@ -1,4 +1,3 @@
 #![allow(unused_variables)]
 pub mod context;
-mod epoch_observer;
 mod utils;
